No commit message
svn path=/trunk/iCub/; revision=15697

+++ src/modules/objectsPropertiesCollector/main.cpp
@@ -1186,7 +1186,7 @@ public:
             unsigned int calls=nCalls-nCallsOld;
             double timeSpent=cumTime-cumTimeOld;
             fprintf(stdout,"*** Statistics: received %d/%g [requests/s]; %g [us/request] spent on average\n",
-                    calls,getPeriod(),timeSpent==0.0?0.0:1e6*(timeSpent/calls));
+                    calls,getPeriod(),timeSpent==0.0?0.0:(1e6*timeSpent)/(double)calls);
 
             nCallsOld=nCalls;
             cumTimeOld=cumTime;
